:recycle: Refactor: Symbol client 위치 수정 및 config 설정 추가
<body>
- 조금 더 읽기 편하게 구조 변경
- symbol을 위한 추가 config 파일 생성

<footer>
- 관련: #86
