+++ clients/client-api/src/main/kotlin/io/say/better/client/config/OpenFeignConfig.kt
@@ -1,0 +1,8 @@
+package io.say.better.client.config
+
+import org.springframework.cloud.openfeign.EnableFeignClients
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+@EnableFeignClients(basePackages = ["io.say.better.client"])
+class OpenFeignConfig
