+++ clients/client-api/src/main/kotlin/io/say/better/client/symbol/client/RecommendClient.kt
@@ -1,7 +1,8 @@
-package io.say.better.client.symbol.api
+package io.say.better.client.symbol.client
 
-import io.say.better.client.symbol.api.dto.RecommendRequest
-import io.say.better.client.symbol.api.dto.RecommendResult
+import io.say.better.client.symbol.api.RecommendApi
+import io.say.better.client.symbol.dto.RecommendRequest
+import io.say.better.client.symbol.dto.RecommendResult
 import io.say.better.client.symbol.converter.RecommendResultConverter
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.PropertySource
