--- clients/client-api/src/main/kotlin/io/say/better/client/symbol/config/OpenFeignConfig.kt
@@ -1,8 +1,0 @@
-package io.say.better.client.symbol.config
-
-import org.springframework.cloud.openfeign.EnableFeignClients
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-@EnableFeignClients(basePackages = ["io.say.better.client.symbol.api"])
-open class OpenFeignConfig
