+++ clients/client-api/src/main/kotlin/io/say/better/client/symbol/dto/RecommendResult.kt
@@ -1,4 +1,4 @@
-package io.say.better.client.symbol.api.dto
+package io.say.better.client.symbol.dto
 
 class RecommendResult private constructor() {
     init {
